Xstream--转换器，只转换特定的字段

+++ src/main/java/com/cxh/converter/XStreamEmptyTagConverter.java
@@ -1,5 +1,6 @@
 package com.cxh.converter;
 
+import java.math.BigDecimal;
 import com.thoughtworks.xstream.converters.Converter;
 import com.thoughtworks.xstream.converters.MarshallingContext;
 import com.thoughtworks.xstream.converters.UnmarshallingContext;
@@ -8,8 +9,9 @@ import com.thoughtworks.xstream.io.HierarchicalStreamWriter;
 
 /**
  * 转换器
- * 将空标签转换成空值
- * 注意，如果空标签对应的是基本类型的字段，则完全不可用，你得返回一个默认的值
+ * 在XML转实体类时，将XMl中为数值类型的空标签，序列成NULL值
+ * 如若不然，会爆数值转换异常。
+ * 注意，如果空标签对应的是基本类型的字段，则此转换器完全不可用，你得返回一个默认的值
  * @author Mr-hang
  *
  */
@@ -17,7 +19,7 @@ public class XStreamEmptyTagConverter implements Converter{
 
 	@Override
 	public boolean canConvert(Class type) {
-		return type.equals(Integer.class);
+		return type.equals(BigDecimal.class);
 	}
 
 	@Override
